Move parseCommand and parseStack to utilities
