+++ src/main.js
@@ -2,7 +2,7 @@ import './polyfills/index';
 import { attach, detach, disable, enable, onAttach, onDetach } from './api';
 
 /**
- * Enable FTwelve and return public API
+ * Enable F-Twelve and return public API
  */
 enable({ show: false });
 export default Object.freeze({
